# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Crousel Ecommerce API
  description: API autentikasi customer/admin, pengaturan toko, katalog, dan notifikasi untuk aplikasi ecommerce.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /profile:
    get:
      tags: [Profile]
      summary: Ambil profil pengguna yang sedang login
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil pengguna
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
    patch:
      tags: [Profile]
      summary: Perbarui profil pengguna
      description: Semua field bersifat opsional; hanya field yang dikirim yang diperbarui. Kirim null untuk mengosongkan field profil.
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              name: Budi Santoso
              whatsapp: '081234567890'
              birth_date: '1995-08-17'
              gender: male
              province: Jawa Barat
              regency: Kota Bandung
              district: Coblong
              village: Dago
              postal_code: '40135'
              address: Jl. Contoh No. 10
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateProfileMultipartRequest'
      responses:
        '200':
          description: Profil berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [Profile]
      summary: Perbarui profil dengan foto profil
      description: Gunakan multipart/form-data untuk memperbarui field profil dan mengunggah foto profil. Endpoint ini setara dengan PATCH /profile untuk mendukung parser multipart Laravel.
      operationId: updateProfileWithAvatar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateProfileMultipartRequest'
      responses:
        '200':
          description: Profil berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /products:
    get:
      tags: [Products]
      summary: Ambil semua produk
      operationId: listProducts
      parameters:
        - name: category
          in: query
          required: false
          schema: { type: string, enum: [Slide, Slop, Wedges] }
          description: Filter kategori produk.
          example: Slide
      responses:
        '200':
          description: Daftar produk
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Product' }
    post:
      tags: [Products]
      summary: Tambah produk
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/ProductRequest' }
            example:
              slug: invoker-full-black
              name: Invoker Full Black
              category: Slide
              target: Women
              color: Hitam, Putih
              availableColors: [Hitam, Putih]
              images: [invoker-front.jpg, invoker-side.jpg]
              description: Sandal slide Invoker dengan desain minimalis full black.
              price: Rp249.000
              features: [Hand Made Crafting, Crousel Embossed Logo]
              availableSizes: ['36', '37', '38', '39', '40']
              availability: Tersedia
      responses:
        '201':
          description: Produk berhasil dibuat
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /products/{product}:
    get:
      tags: [Products]
      summary: Ambil detail produk berdasarkan ID
      operationId: getProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Detail produk
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Products]
      summary: Perbarui produk berdasarkan ID
      description: Gunakan `multipart/form-data` untuk field produk dan file gambar. Jika client tidak dapat mengirim file dengan metode PATCH, kirim POST ke URL yang sama dengan field `_method=PATCH`.
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/ProductUpdateRequest' }
      responses:
        '200':
          description: Produk berhasil diperbarui
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Products]
      summary: Hapus produk berdasarkan ID
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Produk berhasil dihapus
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
              example: { message: Produk berhasil dihapus. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
      description: Gunakan token dari endpoint register/login sebagai `Bearer <token>`.
  parameters:
    ProductId:
      name: product
      in: path
      required: true
      description: ID produk
      schema: { type: integer, format: int64, example: 1 }
...
  schemas:
    User:
      type: object
      required: [id, name, whatsapp, birth_date, gender, province, regency, district, village, postal_code, address, created_at, updated_at]
      properties:
        id: { type: integer, format: int64, example: 1 }
        name: { type: string, maxLength: 100, example: Budi Santoso }
        whatsapp: { type: string, description: 'Nomor disimpan dalam format internasional tanpa tanda plus', pattern: '^62[0-9]{8,13}$', example: '6281234567890' }
        birth_date: { type: string, format: date, nullable: true, example: '1995-08-17' }
        gender: { type: string, enum: [male, female, other], nullable: true, example: male }
        province: { type: string, nullable: true, example: Jawa Barat }
        province_id: { type: integer, nullable: true, example: 9 }
        regency: { type: string, nullable: true, example: Kota Bandung }
        regency_id: { type: integer, nullable: true, example: 108 }
        district: { type: string, nullable: true, example: Coblong }
        district_id: { type: integer, nullable: true, example: 1543 }
        village: { type: string, nullable: true, example: Dago }
        village_id: { type: integer, nullable: true, example: 22550 }
        postal_code: { type: string, nullable: true, example: '40135' }
        address: { type: string, nullable: true, example: Jl. Contoh No. 10 }
        avatar_url: { type: string, format: uri, nullable: true, readOnly: true, description: URL publik foto profil }
        created_at: { type: string, format: date-time, example: '2026-09-05T13:00:00.000000Z' }
        updated_at: { type: string, format: date-time, example: '2026-09-05T13:00:00.000000Z' }
...
    UpdateProfileRequest:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        whatsapp: { type: string, example: '081234567890' }
        birth_date: { type: string, format: date, nullable: true }
        gender: { type: string, enum: [male, female, other], nullable: true }
        province: { type: string, maxLength: 100, nullable: true }
        province_id: { type: integer, minimum: 1, nullable: true }
        regency: { type: string, maxLength: 100, nullable: true }
        regency_id: { type: integer, minimum: 1, nullable: true }
        district: { type: string, maxLength: 100, nullable: true }
        district_id: { type: integer, minimum: 1, nullable: true }
        village: { type: string, maxLength: 100, nullable: true }
        village_id: { type: integer, minimum: 1, nullable: true }
        postal_code: { type: string, maxLength: 10, nullable: true }
        address: { type: string, maxLength: 500, nullable: true }
    UpdateProfileMultipartRequest:
      type: object
      description: Field profil opsional. Gunakan multipart/form-data saat mengunggah foto profil.
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        whatsapp: { type: string, example: '081234567890' }
        birth_date: { type: string, format: date, nullable: true }
        gender: { type: string, enum: [male, female, other], nullable: true }
        province: { type: string, maxLength: 100, nullable: true }
        province_id: { type: integer, minimum: 1, nullable: true }
        regency: { type: string, maxLength: 100, nullable: true }
        regency_id: { type: integer, minimum: 1, nullable: true }
        district: { type: string, maxLength: 100, nullable: true }
        district_id: { type: integer, minimum: 1, nullable: true }
        village: { type: string, maxLength: 100, nullable: true }
        village_id: { type: integer, minimum: 1, nullable: true }
        postal_code: { type: string, maxLength: 10, nullable: true }
        address: { type: string, maxLength: 500, nullable: true }
        avatar: { type: string, format: binary, description: 'Foto profil JPEG, PNG, WebP, atau GIF maksimal 5 MB' }
...
    ProfileResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            user: { $ref: '#/components/schemas/User' }
...
    Product:
      type: object
      required: [id, slug, name, category, target, color, availableColors, images, description, price, weight, features, availableSizes, availability, created_at, updated_at]
      properties:
        id: { type: integer, format: int64, example: 1 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', example: invoker-full-black }
        name: { type: string, example: Invoker Full Black }
        category: { type: string, enum: [Slide, Slop, Wedges], example: Slide }
        target: { type: string, enum: [Women, Men, Unisex, Kids], example: Women }
        color: { type: string, description: 'Satu atau beberapa warna dipisahkan koma', example: 'Hitam, Putih' }
        availableColors: { type: array, items: { type: string }, description: 'Subset warna dari field color yang tersedia untuk dipilih pembeli', example: [Hitam, Putih] }
        images: { type: array, items: { type: string, format: uri }, description: 'URL publik gambar yang dikembalikan API.', example: [http://localhost:8000/storage/products/invoker-front.jpg] }
        description: { type: string, example: Sandal slide Invoker dengan desain minimalis full black. }
        price: { type: string, example: Rp249.000 }
        weight: { type: integer, minimum: 1, description: 'Berat produk dalam gram.', example: 500 }
        features: { type: array, items: { type: string }, example: [Hand Made Crafting, Crousel Embossed Logo] }
        availableSizes: { type: array, items: { type: string }, example: ['36', '37', '38', '39', '40'] }
        availability: { type: string, enum: [Tersedia, Pre-order, Habis], example: Tersedia }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ProductRequest:
      allOf:
        - $ref: '#/components/schemas/ProductCreateFields'
    ProductUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/ProductUpdateFields'
...
    ProductCreateFields:
      type: object
      required: [slug, name, category, target, color, availableColors, images, description, price, weight, features, availableSizes, availability]
      properties:
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        name: { type: string, maxLength: 150 }
        category: { type: string, enum: [Slide, Slop, Wedges] }
        target: { type: string, enum: [Women, Men, Unisex, Kids] }
        color: { type: string, maxLength: 255, description: 'Satu atau beberapa warna dipisahkan koma' }
        availableColors: { type: array, minItems: 1, items: { type: string } }
        images: { type: array, minItems: 1, items: { type: string, format: binary }, description: 'Satu atau beberapa file gambar (JPEG, PNG, WebP, atau GIF), maksimal 5 MB per file.' }
        description: { type: string }
        price: { type: string, maxLength: 50 }
        weight: { type: integer, minimum: 1, description: 'Berat produk dalam gram.' }
        features: { type: array, items: { type: string } }
        availableSizes: { type: array, minItems: 1, items: { type: string } }
        availability: { type: string, enum: [Tersedia, Pre-order, Habis] }
    ProductUpdateFields:
      type: object
      description: Semua field opsional; hanya field yang dikirim akan diperbarui.
      properties:
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        name: { type: string, maxLength: 150 }
        category: { type: string, enum: [Slide, Slop, Wedges] }
        target: { type: string, enum: [Women, Men, Unisex, Kids] }
        color: { type: string, maxLength: 255 }
        availableColors: { type: array, items: { type: string } }
        images: { type: array, items: { type: string, format: binary }, description: 'Opsional. Jika dikirim, gambar baru akan ditambahkan (append) ke daftar gambar saat ini. Maksimal 5 MB per file.' }
        deletedImages: { type: array, items: { type: string, format: uri }, description: 'Daftar URL gambar yang sudah ada (dari respons API) untuk dihapus dari produk ini.' }
        description: { type: string }
        price: { type: string, maxLength: 50 }
        weight: { type: integer, minimum: 1, description: 'Berat produk dalam gram.' }
        features: { type: array, items: { type: string } }
        availableSizes: { type: array, items: { type: string } }
        availability: { type: string, enum: [Tersedia, Pre-order, Habis] }
    ProductResponse:
      type: object
      required: [data]
      properties:
        message: { type: string }
        data:
          type: object
          properties:
            product: { $ref: '#/components/schemas/Product' }
    MessageResponse:
      type: object
      properties:
        message: { type: string }
    ValidationError:
      type: object
      properties:
        message: { type: string, example: The given data was invalid. }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
          example:
            whatsapp: [The whatsapp has already been taken.]
    UnauthenticatedError:
      type: object
      properties:
        message: { type: string, example: Unauthenticated. }
  responses:
    ValidationError:
      description: Data tidak valid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
    Unauthenticated:
      description: Token tidak ada atau tidak valid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UnauthenticatedError' }
    NotFound:
      description: Resource tidak ditemukan
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string, example: 'No query results for model [App\\Models\\Product] 99.' }